--- src/components/TextField/index.ts
@@ -1,1 +1,0 @@
-export * from './TextField';
