--- src/stories/TextField.stories.tsx
@@ -1,16 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react';
-import { TextField } from '@mui/material';
-
-const meta: Meta<typeof TextField> = {
-  title: 'TextField',
-  component: TextField,
-  tags: ['autodocs'],
-};
-
-export default meta;
-
-type Story = StoryObj<typeof TextField>;
-
-export const MIUTextFiled: Story = {
-  render: () => <TextField variant="standard" label="label" />,
-};
